Initial Commit
Eventually this addon will allow for the exporting of voxel models from blender.

First select the mesh you wish to work with, remesh it via the below steps, and bring the mesh into edit mode.

Currently the code will select all faces selecting an axial direction,
and will scale voxels to one blender unit a piece.

For now the mesh must be manually voxelized via:
	Modifiers
	Remesh
	Mode = Blocks

The search axis is also hard coded for now, axis = 0 is the x direction,
axis = 1 is the y direction, and the axis = 2 is the z direction.